Avoid unsigned data types

+++ compute_samples/applications/commands_aggregation/include/commands_aggregation/commands_aggregation.hpp
@@ -36,10 +36,9 @@ namespace compute = boost::compute;
 namespace compute_samples {
 class CommandsAggregationApplication : public Application {
 public:
-  std::vector<uint32_t>
-  run_workloads_out_of_order(const size_t global_work_size,
-                             src::logger &logger) const;
-  std::vector<uint32_t> run_workloads_in_order(const size_t global_work_size,
+  std::vector<uint32_t> run_workloads_out_of_order(const int global_work_size,
+                                                   src::logger &logger) const;
+  std::vector<uint32_t> run_workloads_in_order(const int global_work_size,
                                                src::logger &logger) const;
 
 private:
@@ -48,7 +47,7 @@ private:
   struct Arguments {
     bool help = false;
     bool in_order = false;
-    size_t global_work_size = 0;
+    int global_work_size = 0;
   };
   Arguments
   parse_command_line(const std::vector<std::string> &command_line) const;
